pyproject.toml
[project]
name = "langchain-mcp-adapters-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.10",
    "langgraph>=0.6.7",
    "mcp[cli]>=1.14.1",
    "python-dotenv>=1.1.1",
]
